+++ p2panda-net/src/sync/log_sync/api.rs
@@ -4,7 +4,7 @@ use std::fmt::Debug;
 use std::marker::PhantomData;
 use std::sync::Arc;
 
-use p2panda_core::{Extensions, Hash, LogId, Operation, Topic, VerifyingKey};
+use p2panda_core::{Extensions, Hash, LogId, Operation, SeqNum, Topic, VerifyingKey};
 use p2panda_store::logs::LogStore;
 use p2panda_store::topics::TopicStore;
 use p2panda_sync::protocols::TopicLogSyncEvent;
@@ -40,7 +40,7 @@ use crate::sync::log_sync::Builder;
 #[derive(Clone, Debug)]
 pub struct LogSync<S, L, E>
 where
-    S: LogStore<Operation<E>, VerifyingKey, L, u64, Hash>
+    S: LogStore<Operation<E>, VerifyingKey, L, SeqNum, Hash>
         + TopicStore<Topic, VerifyingKey, L>
         + Clone
         + Send
@@ -63,7 +63,7 @@ where
 
 impl<S, L, E> LogSync<S, L, E>
 where
-    S: LogStore<Operation<E>, VerifyingKey, L, u64, Hash>
+    S: LogStore<Operation<E>, VerifyingKey, L, SeqNum, Hash>
         + TopicStore<Topic, VerifyingKey, L>
         + Clone
         + Send